requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-mgmt-resource==21.2.1
azure-mgmt-compute==29.0.0
azure-mgmt-network==22.1.0
azure-identity==1.12.0
azure-common==1.1.28
msrestazure==0.6.4
azure-cli-core==2.42.0